separate config objects into multiple files
